Changed msg.Message.Amount/StockId to uint64
This means that every field in message, apart from Kind, TradeId and
TraderId, are uint64. This simplifies the code that deals with these
values. In particular this was motivated by the fact that the total
value of a BUY/SELL message is Amount * Price, previously Price was
uint64 and Amount uint32 making for awkward arithmetic.

+++ client/data.go
@@ -18,7 +18,7 @@ type Response struct {
 type traderState struct {
 	CurrentBalance   uint64            `json:"currentBalance"`
 	AvailableBalance uint64            `json:"availableBalance"`
-	StocksHeld       map[string]uint32 `json:"stocksHeld"`
-	StocksToSell     map[string]uint32 `json:"stocksToSell"`
+	StocksHeld       map[string]uint64 `json:"stocksHeld"`
+	StocksToSell     map[string]uint64 `json:"stocksToSell"`
 	Outstanding      []msg.Message     `json:"outstanding"`
 }
