+++ client/ttrader_unit_test.go
@@ -19,26 +19,26 @@ func TestMessageProcessBuyCancelCancelled(t *testing.T) {
 	traderId := uint32(1)
 	tdr := setupTrader(traderId, t)
 	expectBalance(t, tdr.balance, 100, 100) // This test expects that initialBalance is 100
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	// Submit buys
 	m1 := &msg.Message{Kind: msg.BUY, TraderId: traderId, TradeId: 0, StockId: 1, Price: 5, Amount: 5}
 	tdr.process(m1)
 	expectBalance(t, tdr.balance, 75, 100)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 	// Cancel
 	tdr.process(&msg.Message{Kind: msg.CANCEL, TraderId: traderId, TradeId: m1.TradeId, StockId: 1, Price: 5, Amount: 5})
 	expectBalance(t, tdr.balance, 75, 100)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 	// Confirm CANCELLED
 	tdr.process(&msg.Message{Kind: msg.CANCELLED, TraderId: traderId, TradeId: m1.TradeId, StockId: 1, Price: 5, Amount: 5})
 	expectBalance(t, tdr.balance, 100, 100)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 }
 
@@ -46,20 +46,20 @@ func TestMessageProcessBuyFullSimple(t *testing.T) {
 	traderId := uint32(1)
 	tdr := setupTrader(traderId, t)
 	expectBalance(t, tdr.balance, 100, 100) // This test expects that initialBalance is 100
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	// Submit buy
 	m1 := &msg.Message{Kind: msg.BUY, TraderId: traderId, TradeId: 0, StockId: 1, Price: 5, Amount: 5}
 	tdr.process(m1)
 	expectBalance(t, tdr.balance, 75, 100)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 	// Full Match
 	tdr.process(&msg.Message{Kind: msg.FULL, TraderId: traderId, TradeId: m1.TradeId, StockId: 1, Price: 5, Amount: 5})
 	expectBalance(t, tdr.balance, 75, 75)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 15, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 15, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 }
 
@@ -67,20 +67,20 @@ func TestMessageProcessBuyFullDiffSimple(t *testing.T) {
 	traderId := uint32(1)
 	tdr := setupTrader(traderId, t)
 	expectBalance(t, tdr.balance, 100, 100) // This test expects that initialBalance is 100
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	// Submit buy
 	m1 := &msg.Message{Kind: msg.BUY, TraderId: traderId, TradeId: 0, StockId: 1, Price: 5, Amount: 5}
 	tdr.process(m1)
 	expectBalance(t, tdr.balance, 75, 100)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 	// Full Match - at lower than bid price
 	tdr.process(&msg.Message{Kind: msg.FULL, TraderId: traderId, TradeId: m1.TradeId, StockId: 1, Price: 4, Amount: 5})
 	expectBalance(t, tdr.balance, 80, 80)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 15, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 15, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 }
 
@@ -88,25 +88,25 @@ func TestMessageProcessBuyPartialSimple(t *testing.T) {
 	traderId := uint32(1)
 	tdr := setupTrader(traderId, t)
 	expectBalance(t, tdr.balance, 100, 100) // This test expects that initialBalance is 100
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	// Submit some buys
 	m1 := &msg.Message{Kind: msg.BUY, TraderId: traderId, TradeId: 0, StockId: 1, Price: 5, Amount: 5}
 	tdr.process(m1)
 	expectBalance(t, tdr.balance, 75, 100)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 	// Partial Match
 	tdr.process(&msg.Message{Kind: msg.PARTIAL, TraderId: traderId, TradeId: m1.TradeId, StockId: 1, Price: 5, Amount: 2})
 	expectBalance(t, tdr.balance, 75, 90)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 12, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 12, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 	tdr.process(&msg.Message{Kind: msg.PARTIAL, TraderId: traderId, TradeId: m1.TradeId, StockId: 1, Price: 5, Amount: 3})
 	expectBalance(t, tdr.balance, 75, 75)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 15, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 15, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 }
 
@@ -114,25 +114,25 @@ func TestMessageProcessBuyPartialDiffSimple(t *testing.T) {
 	traderId := uint32(1)
 	tdr := setupTrader(traderId, t)
 	expectBalance(t, tdr.balance, 100, 100) // This test expects that initialBalance is 100
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	// Submit some buys
 	m1 := &msg.Message{Kind: msg.BUY, TraderId: traderId, TradeId: 0, StockId: 1, Price: 5, Amount: 5}
 	tdr.process(m1)
 	expectBalance(t, tdr.balance, 75, 100)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 	// Partial Matches at lower than bid price
 	tdr.process(&msg.Message{Kind: msg.PARTIAL, TraderId: traderId, TradeId: m1.TradeId, StockId: 1, Price: 4, Amount: 2})
 	expectBalance(t, tdr.balance, 77, 92)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 12, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 12, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 	tdr.process(&msg.Message{Kind: msg.PARTIAL, TraderId: traderId, TradeId: m1.TradeId, StockId: 1, Price: 3, Amount: 3})
 	expectBalance(t, tdr.balance, 83, 83)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 15, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 15, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 }
 
@@ -140,26 +140,26 @@ func TestMessageProcessSellCancelCancelled(t *testing.T) {
 	traderId := uint32(1)
 	tdr := setupTrader(traderId, t)
 	expectBalance(t, tdr.balance, 100, 100) // This test expects that initialBalance is 100
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	// Submit sell
 	m1 := &msg.Message{Kind: msg.SELL, TraderId: traderId, TradeId: 0, StockId: 1, Price: 5, Amount: 5}
 	tdr.process(m1)
 	expectBalance(t, tdr.balance, 100, 100)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 5, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{1: 5})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 5, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{1: 5})
 	validate(t, tdr)
 	// Cancel sell
 	tdr.process(&msg.Message{Kind: msg.CANCEL, TraderId: traderId, TradeId: m1.TradeId, StockId: 1, Price: 5, Amount: 5})
 	expectBalance(t, tdr.balance, 100, 100)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 5, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{1: 5})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 5, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{1: 5})
 	validate(t, tdr)
 	// Confirm CANCELLED
 	tdr.process(&msg.Message{Kind: msg.CANCELLED, TraderId: traderId, TradeId: m1.TradeId, StockId: 1, Price: 5, Amount: 5})
 	expectBalance(t, tdr.balance, 100, 100)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 }
 
@@ -167,20 +167,20 @@ func TestMessageProcessSellFullSimple(t *testing.T) {
 	traderId := uint32(1)
 	tdr := setupTrader(traderId, t)
 	expectBalance(t, tdr.balance, 100, 100) // This test expects that initialBalance is 100
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	// Submit sell
 	m1 := &msg.Message{Kind: msg.SELL, TraderId: traderId, TradeId: 0, StockId: 1, Price: 5, Amount: 5}
 	tdr.process(m1)
 	expectBalance(t, tdr.balance, 100, 100)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 5, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{1: 5})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 5, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{1: 5})
 	validate(t, tdr)
 	// Match sell FULL
 	tdr.process(&msg.Message{Kind: msg.FULL, TraderId: traderId, TradeId: m1.TradeId, StockId: 1, Price: 5, Amount: 5})
 	expectBalance(t, tdr.balance, 125, 125)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 5, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 5, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 }
 
@@ -188,20 +188,20 @@ func TestMessageProcessSellFullDiff(t *testing.T) {
 	traderId := uint32(1)
 	tdr := setupTrader(traderId, t)
 	expectBalance(t, tdr.balance, 100, 100) // This test expects that initialBalance is 100
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	// Submit sell
 	m1 := &msg.Message{Kind: msg.SELL, TraderId: traderId, TradeId: 0, StockId: 1, Price: 5, Amount: 5}
 	tdr.process(m1)
 	expectBalance(t, tdr.balance, 100, 100)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 5, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{1: 5})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 5, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{1: 5})
 	validate(t, tdr)
 	// Match sell FUll
 	tdr.process(&msg.Message{Kind: msg.FULL, TraderId: traderId, TradeId: m1.TradeId, StockId: 1, Price: 7, Amount: 5})
 	expectBalance(t, tdr.balance, 135, 135)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 5, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 5, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	validate(t, tdr)
 }
 
@@ -209,20 +209,20 @@ func TestMessageProcessSellPartialSimple(t *testing.T) {
 	traderId := uint32(1)
 	tdr := setupTrader(traderId, t)
 	expectBalance(t, tdr.balance, 100, 100) // This test expects that initialBalance is 100
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	// Submit sell
 	m1 := &msg.Message{Kind: msg.SELL, TraderId: traderId, TradeId: 0, StockId: 1, Price: 5, Amount: 5}
 	tdr.process(m1)
 	expectBalance(t, tdr.balance, 100, 100)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 5, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{1: 5})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 5, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{1: 5})
 	validate(t, tdr)
 	// Match sell PARTIAL
 	tdr.process(&msg.Message{Kind: msg.PARTIAL, TraderId: traderId, TradeId: m1.TradeId, StockId: 1, Price: 5, Amount: 3})
 	expectBalance(t, tdr.balance, 115, 115)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 5, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{1: 2})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 5, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{1: 2})
 	validate(t, tdr)
 }
 
@@ -230,20 +230,20 @@ func TestMessageProcessSellPartialDiff(t *testing.T) {
 	traderId := uint32(1)
 	tdr := setupTrader(traderId, t)
 	expectBalance(t, tdr.balance, 100, 100) // This test expects that initialBalance is 100
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 10, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 10, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{})
 	// Submit sell
 	m1 := &msg.Message{Kind: msg.SELL, TraderId: traderId, TradeId: 0, StockId: 1, Price: 5, Amount: 5}
 	tdr.process(m1)
 	expectBalance(t, tdr.balance, 100, 100)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 5, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{1: 5})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 5, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{1: 5})
 	validate(t, tdr)
 	// Match sell PARTIAL
 	tdr.process(&msg.Message{Kind: msg.PARTIAL, TraderId: traderId, TradeId: m1.TradeId, StockId: 1, Price: 7, Amount: 3})
 	expectBalance(t, tdr.balance, 121, 121)
-	expectInMap(t, tdr.stocks.held, map[uint32]uint32{1: 5, 2: 10, 3: 10})
-	expectInMap(t, tdr.stocks.toSell, map[uint32]uint32{1: 2})
+	expectInMap(t, tdr.stocks.held, map[uint64]uint64{1: 5, 2: 10, 3: 10})
+	expectInMap(t, tdr.stocks.toSell, map[uint64]uint64{1: 2})
 	validate(t, tdr)
 }
